🔨 Moving date parsing to util
